refactor: moved RenameRequest task_1110
